requirements.txt
pandas>=2.0.0
Pillow>=10.2.0
qdrant-client>=1.9.0
requests>=2.31.0
sentence-transformers>=2.2.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
openai>=1.0.0
groq>=0.4.0
numpy>=1.24.0